fix incorrect player_facing_direction

+++ src/Hardware/IR/GameData.h
@@ -8,7 +8,7 @@ class GameData
 public:
     uint16_t player_x_position;
     uint8_t player_y_position;
-    uint16_t player_facing_direction;
+    double player_facing_direction;
     bool player_shot_bullet;
     bool player_died;
     bool wave_ended;
